Extract `SyncPendingProvider` from `SyncFrameworkIntegration`
