+++ core/src/main/kotlin/at/bitfire/davdroid/sync/adapter/SyncFrameworkIntegration.kt
@@ -6,23 +6,12 @@ package at.bitfire.davdroid.sync.adapter
 
 import android.accounts.Account
 import android.content.ContentResolver
-import android.content.Context
 import androidx.annotation.WorkerThread
-import at.bitfire.davdroid.accounts.AccountId
-import at.bitfire.davdroid.accounts.AndroidAccountManager
-import at.bitfire.davdroid.accounts.toAndroidAccount
-import at.bitfire.davdroid.resource.LocalAddressBookStore
-import at.bitfire.davdroid.sync.SyncDataType
-import dagger.Lazy
-import dagger.hilt.android.qualifiers.ApplicationContext
 import kotlinx.coroutines.ExperimentalCoroutinesApi
 import kotlinx.coroutines.channels.awaitClose
 import kotlinx.coroutines.flow.Flow
 import kotlinx.coroutines.flow.callbackFlow
-import kotlinx.coroutines.flow.combine
-import kotlinx.coroutines.flow.distinctUntilChanged
 import kotlinx.coroutines.flow.flatMapLatest
-import kotlinx.coroutines.flow.flowOf
 import java.util.logging.Logger
 import javax.inject.Inject
 
@@ -34,9 +23,6 @@ import javax.inject.Inject
  * Sync requests from the Sync Adapter Framework are handled by [SyncAdapterService].
  */
 class SyncFrameworkIntegration @Inject constructor(
-    private val androidAccountManager: AndroidAccountManager,
-    @ApplicationContext private val context: Context,
-    private val localAddressBookStore: Lazy<LocalAddressBookStore>,
     private val logger: Logger
 ) {
 
@@ -173,45 +159,6 @@ class SyncFrameworkIntegration @Inject constructor(
     }
 
     /**
-     * Observe whether any of the given data types is currently pending for sync.
-     *
-     * _Note:_ the sync framework doesn't reliably mark a finished one-time sync as "not pending"
-     * anymore, see [SyncAdapterImpl.hasAlwaysPendingIssue].
-     *
-     * @param accountId [AccountId] of the account to observe sync status for
-     * @param dataTypes data types to observe sync status for
-     *
-     * @return flow emitting true if any of the given data types has a sync pending, false otherwise
-     */
-    @OptIn(ExperimentalCoroutinesApi::class)
-    fun isSyncPending(accountId: AccountId, dataTypes: Iterable<SyncDataType>): Flow<Boolean> {
-        // Determine the pending state for each data type of the account as separate flows
-        val pendingStateFlows: List<Flow<Boolean>> = dataTypes.mapNotNull { dataType ->
-            // Map datatype to authority
-            dataType.currentAuthority(context)?.let { authority ->
-                // If checking contacts, we need to check all address book accounts instead of the single main account
-                val accountsFlow: Flow<List<Account>> = when (dataType) {
-                    SyncDataType.CONTACTS -> {
-                        localAddressBookStore.get().getAddressBookAccountsFlow(accountId.toAndroidAccount())
-                    }
-                    else -> {
-                        val account = androidAccountManager.getAndroidAccount(accountId)
-                        flowOf(listOf(account))
-                    }
-                }
-
-                // Return the pending state flow for accounts with this authority
-                anyPendingSyncFlow(accountsFlow, authority)
-            }
-        }
-
-        // Combine the different per data type pending state flows into one
-        return combine(pendingStateFlows) { pendingStates ->
-            pendingStates.any { pending -> pending }
-        }.distinctUntilChanged()
-    }
-
-    /**
      * Maps the given accounts flow to a simple boolean flow telling us whether any of the accounts
      * has a pending sync for given authority.
      *
@@ -222,7 +169,7 @@ class SyncFrameworkIntegration @Inject constructor(
      * the given authority and *false* otherwise
      */
     @OptIn(ExperimentalCoroutinesApi::class)
-    private fun anyPendingSyncFlow(
+    fun anyPendingSyncFlow(
         accountsFlow: Flow<List<Account>>,
         authority: String
     ): Flow<Boolean> = accountsFlow.flatMapLatest { accounts ->
